+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 v0.9.6
 ^^^^^^
 
+* Added web analytics (umami)
 * Added a data export button
 * Added an account deletion button
 * Added a Render demo deployment